requirements.txt
joblib==1.1.0
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
scikit_learn==1.1.2
streamlit==1.11.1
